recompiler:  Replace (unsigned long) with (u32).

+++ r4300/x86/gbc.c
@@ -36,22 +36,22 @@
 
 void genbc1f_test()
 {
-   test_m32_imm32((unsigned long*)&FCR31, 0x800000);
+   test_m32_imm32((u32*)&FCR31, 0x800000);
    jne_rj(12);
-   mov_m32_imm32((unsigned long*)(&branch_taken), 1); // 10
+   mov_m32_imm32((u32*)(&branch_taken), 1); // 10
    jmp_imm_short(10); // 2
-   mov_m32_imm32((unsigned long*)(&branch_taken), 0); // 10
+   mov_m32_imm32((u32*)(&branch_taken), 0); // 10
 }
 
 void genbc1f()
 {
 #ifdef INTERPRET_BC1F
-   gencallinterp((unsigned long)BC1F, 1);
+   gencallinterp((u32)BC1F, 1);
 #else
    if (((dst->addr & 0xFFF) == 0xFFC &&
        (dst->addr < 0x80000000 || dst->addr >= 0xC0000000))||no_compiled_jump)
      {
-	gencallinterp((unsigned long)BC1F, 1);
+	gencallinterp((u32)BC1F, 1);
 	return;
      }
    
@@ -65,12 +65,12 @@ void genbc1f()
 void genbc1f_out()
 {
 #ifdef INTERPRET_BC1F_OUT
-   gencallinterp((unsigned long)BC1F_OUT, 1);
+   gencallinterp((u32)BC1F_OUT, 1);
 #else
    if (((dst->addr & 0xFFF) == 0xFFC &&
        (dst->addr < 0x80000000 || dst->addr >= 0xC0000000))||no_compiled_jump)
      {
-	gencallinterp((unsigned long)BC1F_OUT, 1);
+	gencallinterp((u32)BC1F_OUT, 1);
 	return;
      }
    
@@ -84,12 +84,12 @@ void genbc1f_out()
 void genbc1f_idle()
 {
 #ifdef INTERPRET_BC1F_IDLE
-   gencallinterp((unsigned long)BC1F_IDLE, 1);
+   gencallinterp((u32)BC1F_IDLE, 1);
 #else
    if (((dst->addr & 0xFFF) == 0xFFC &&
        (dst->addr < 0x80000000 || dst->addr >= 0xC0000000))||no_compiled_jump)
      {
-	gencallinterp((unsigned long)BC1F_IDLE, 1);
+	gencallinterp((u32)BC1F_IDLE, 1);
 	return;
      }
    
@@ -102,22 +102,22 @@ void genbc1f_idle()
 
 void genbc1t_test()
 {
-   test_m32_imm32((unsigned long*)&FCR31, 0x800000);
+   test_m32_imm32((u32*)&FCR31, 0x800000);
    je_rj(12);
-   mov_m32_imm32((unsigned long*)(&branch_taken), 1); // 10
+   mov_m32_imm32((u32*)(&branch_taken), 1); // 10
    jmp_imm_short(10); // 2
-   mov_m32_imm32((unsigned long*)(&branch_taken), 0); // 10
+   mov_m32_imm32((u32*)(&branch_taken), 0); // 10
 }
 
 void genbc1t()
 {
 #ifdef INTERPRET_BC1T
-   gencallinterp((unsigned long)BC1T, 1);
+   gencallinterp((u32)BC1T, 1);
 #else
    if (((dst->addr & 0xFFF) == 0xFFC &&
        (dst->addr < 0x80000000 || dst->addr >= 0xC0000000))||no_compiled_jump)
      {
-	gencallinterp((unsigned long)BC1T, 1);
+	gencallinterp((u32)BC1T, 1);
 	return;
      }
    
@@ -131,12 +131,12 @@ void genbc1t()
 void genbc1t_out()
 {
 #ifdef INTERPRET_BC1T_OUT
-   gencallinterp((unsigned long)BC1T_OUT, 1);
+   gencallinterp((u32)BC1T_OUT, 1);
 #else
    if (((dst->addr & 0xFFF) == 0xFFC &&
        (dst->addr < 0x80000000 || dst->addr >= 0xC0000000))||no_compiled_jump)
      {
-	gencallinterp((unsigned long)BC1T_OUT, 1);
+	gencallinterp((u32)BC1T_OUT, 1);
 	return;
      }
    
@@ -150,12 +150,12 @@ void genbc1t_out()
 void genbc1t_idle()
 {
 #ifdef INTERPRET_BC1T_IDLE
-   gencallinterp((unsigned long)BC1T_IDLE, 1);
+   gencallinterp((u32)BC1T_IDLE, 1);
 #else
    if (((dst->addr & 0xFFF) == 0xFFC &&
        (dst->addr < 0x80000000 || dst->addr >= 0xC0000000))||no_compiled_jump)
      {
-	gencallinterp((unsigned long)BC1T_IDLE, 1);
+	gencallinterp((u32)BC1T_IDLE, 1);
 	return;
      }
    
@@ -169,12 +169,12 @@ void genbc1t_idle()
 void genbc1fl()
 {
 #ifdef INTERPRET_BC1FL
-   gencallinterp((unsigned long)BC1FL, 1);
+   gencallinterp((u32)BC1FL, 1);
 #else
    if (((dst->addr & 0xFFF) == 0xFFC &&
        (dst->addr < 0x80000000 || dst->addr >= 0xC0000000))||no_compiled_jump)
      {
-	gencallinterp((unsigned long)BC1FL, 1);
+	gencallinterp((u32)BC1FL, 1);
 	return;
      }
    
@@ -188,12 +188,12 @@ void genbc1fl()
 void genbc1fl_out()
 {
 #ifdef INTERPRET_BC1FL_OUT
-   gencallinterp((unsigned long)BC1FL_OUT, 1);
+   gencallinterp((u32)BC1FL_OUT, 1);
 #else
    if (((dst->addr & 0xFFF) == 0xFFC &&
        (dst->addr < 0x80000000 || dst->addr >= 0xC0000000))||no_compiled_jump)
      {
-	gencallinterp((unsigned long)BC1FL_OUT, 1);
+	gencallinterp((u32)BC1FL_OUT, 1);
 	return;
      }
    
@@ -207,12 +207,12 @@ void genbc1fl_out()
 void genbc1fl_idle()
 {
 #ifdef INTERPRET_BC1FL_IDLE
-   gencallinterp((unsigned long)BC1FL_IDLE, 1);
+   gencallinterp((u32)BC1FL_IDLE, 1);
 #else
    if (((dst->addr & 0xFFF) == 0xFFC &&
        (dst->addr < 0x80000000 || dst->addr >= 0xC0000000))||no_compiled_jump)
      {
-	gencallinterp((unsigned long)BC1FL_IDLE, 1);
+	gencallinterp((u32)BC1FL_IDLE, 1);
 	return;
      }
    
@@ -226,12 +226,12 @@ void genbc1fl_idle()
 void genbc1tl()
 {
 #ifdef INTERPRET_BC1TL
-   gencallinterp((unsigned long)BC1TL, 1);
+   gencallinterp((u32)BC1TL, 1);
 #else
    if (((dst->addr & 0xFFF) == 0xFFC &&
        (dst->addr < 0x80000000 || dst->addr >= 0xC0000000))||no_compiled_jump)
      {
-	gencallinterp((unsigned long)BC1TL, 1);
+	gencallinterp((u32)BC1TL, 1);
 	return;
      }
    
@@ -245,12 +245,12 @@ void genbc1tl()
 void genbc1tl_out()
 {
 #ifdef INTERPRET_BC1TL_OUT
-   gencallinterp((unsigned long)BC1TL_OUT, 1);
+   gencallinterp((u32)BC1TL_OUT, 1);
 #else
    if (((dst->addr & 0xFFF) == 0xFFC &&
        (dst->addr < 0x80000000 || dst->addr >= 0xC0000000))||no_compiled_jump)
      {
-	gencallinterp((unsigned long)BC1TL_OUT, 1);
+	gencallinterp((u32)BC1TL_OUT, 1);
 	return;
      }
    
@@ -264,12 +264,12 @@ void genbc1tl_out()
 void genbc1tl_idle()
 {
 #ifdef INTERPRET_BC1TL_IDLE
-   gencallinterp((unsigned long)BC1TL_IDLE, 1);
+   gencallinterp((u32)BC1TL_IDLE, 1);
 #else
    if (((dst->addr & 0xFFF) == 0xFFC &&
        (dst->addr < 0x80000000 || dst->addr >= 0xC0000000))||no_compiled_jump)
      {
-	gencallinterp((unsigned long)BC1TL_IDLE, 1);
+	gencallinterp((u32)BC1TL_IDLE, 1);
 	return;
      }
