+++ r4300/x86/gcop1_l.c
@@ -36,12 +36,12 @@
 void gencvt_s_l()
 {
 #ifdef INTERPRET_CVT_S_L
-   gencallinterp((unsigned long)CVT_S_L, 0);
+   gencallinterp((u32)CVT_S_L, 0);
 #else
    gencheck_cop1_unusable();
-   mov_eax_memoffs32((unsigned long*)(&reg_cop1_double[dst->f.cf.fs]));
+   mov_eax_memoffs32((u32*)(&reg_cop1_double[dst->f.cf.fs]));
    fild_preg32_qword(EAX);
-   mov_eax_memoffs32((unsigned long*)(&reg_cop1_simple[dst->f.cf.fd]));
+   mov_eax_memoffs32((u32*)(&reg_cop1_simple[dst->f.cf.fd]));
    fstp_preg32_dword(EAX);
 #endif
 }
@@ -49,12 +49,12 @@ void gencvt_s_l()
 void gencvt_d_l()
 {
 #ifdef INTERPRET_CVT_D_L
-   gencallinterp((unsigned long)CVT_D_L, 0);
+   gencallinterp((u32)CVT_D_L, 0);
 #else
    gencheck_cop1_unusable();
-   mov_eax_memoffs32((unsigned long*)(&reg_cop1_double[dst->f.cf.fs]));
+   mov_eax_memoffs32((u32*)(&reg_cop1_double[dst->f.cf.fs]));
    fild_preg32_qword(EAX);
-   mov_eax_memoffs32((unsigned long*)(&reg_cop1_double[dst->f.cf.fd]));
+   mov_eax_memoffs32((u32*)(&reg_cop1_double[dst->f.cf.fd]));
    fstp_preg32_qword(EAX);
 #endif
 }
